draw: add player paddle to screen
About the offsets, they were added
because without that, the rectangle (paddle)
would otherwise be draw a little lower and
a little to the right than in the middle of
the screen.

This happens because the graphics::draw
function starts drawing a rectangle on the top
left of it, not on the middle.

To calculate the offset, I've divided the height and
width by half, and made it negative, to make the
rectangle be drawn a little higher and to the middle.